2022. 10. 27. STEP3 수정
